Refactor : record파트 service 인터페이스 추가
